Refactor MergeChangesetUtils
Splits MergeChangestUtils into:
- MergeChangesetsImplicits: implicit
  values required for the MergeChangesets
  app to function
- osmesa.apps.DbUtils: Provides more generic
  functions for interacting with a conncted
  PostgreSQL database. These functions are used
  in other command apps.
